refactor!: rename wrapper module to io
BREAKING CHANGE: now wrapper module becomes io.
